requirements.txt
aiohttp==3.9.3
aiopg==1.4.0
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
colored==2.2.4
data-printer==0.0.8
dist-info==0.1.1
frozenlist==1.4.1
idna==3.6
iniconfig==2.0.0
multidict==6.0.5
packaging==23.2
pluggy==1.4.0
psycopg2-binary==2.9.9
pytest==8.0.2
pytest-asyncio==0.23.5
python-dotenv==1.0.1
yarl==1.9.4
